pyproject.toml
[tool.poetry]
name = "cdk-service"
version = "0.1.0"
description = ""
authors = ["myeonggil <echosoul1994@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
aws-cdk-lib = "^2.137.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
